[change] uintからintに変更

+++ models/uint.go
@@ -1,6 +1,6 @@
 package models
 
 type Uint struct {
-	ID   uint8  `gorm:"primary_key"`
+	ID   int  `gorm:"primary_key"`
 	Name string `gorm:"unique"`
 }
